1. Removed the Logger from TestFixture 2. Implemented a BaseClass for better reuse of test code

+++ test/AWS.Logger.TestUtils/TestFixture.cs
@@ -18,11 +18,11 @@ namespace AWS.Logger.TestUtils
     public class TestFixture : IDisposable
     {
         public List<string> LogGroupNameList;
-        public AmazonCloudWatchLogsClient Client;
 
         public TestFixture()
         {
-            Client = new AmazonCloudWatchLogsClient(Amazon.RegionEndpoint.USWest2);
+            AmazonCloudWatchLogsClient Client = 
+                new AmazonCloudWatchLogsClient(Amazon.RegionEndpoint.USWest2);
             LogGroupNameList = new List<string>();
         }
 
@@ -34,57 +34,26 @@ namespace AWS.Logger.TestUtils
 
         protected virtual void Dispose(bool disposing)
         {
-            foreach(var logGroupName in LogGroupNameList)
+            AmazonCloudWatchLogsClient Client = 
+                new AmazonCloudWatchLogsClient(Amazon.RegionEndpoint.USWest2);
+            foreach (var logGroupName in LogGroupNameList)
             {
                 if (!(string.IsNullOrEmpty(logGroupName)))
                 {
-                    var response = Client.DeleteLogGroupAsync(new DeleteLogGroupRequest
+                    DescribeLogGroupsResponse describeLogGroupsResponse = Client.DescribeLogGroupsAsync(
+                    new DescribeLogGroupsRequest
                     {
-                        LogGroupName = logGroupName
-                    }).Result;
-                }
-            }
-        }
-        public bool NotifyLoggingCompleted(string logGroupName, string filterPattern)
-        {
-            Stopwatch timer = new Stopwatch();
-            timer.Start();
-            while (timer.Elapsed.TotalSeconds<10) 
-            {
-                Thread.Sleep(500);
-                if(IsFilterPatternExists(logGroupName, filterPattern))
-                {
-                    break;
-                }
-            }
-            return IsFilterPatternExists(logGroupName, filterPattern);
-        }
-        public bool IsFilterPatternExists(string logGroupName, string filterPattern)
-        {
-            DescribeLogStreamsResponse describeLogstreamsResponse = Client.
-                    DescribeLogStreamsAsync(new DescribeLogStreamsRequest
-                    {
-                        Descending = true,
-                        LogGroupName = logGroupName,
-                        OrderBy = "LastEventTime"
-                    }).Result;
-            if (describeLogstreamsResponse.LogStreams.Count > 0)
-            {
-                List<string> logStreamNames = new List<string>();
-                logStreamNames.Add(describeLogstreamsResponse.LogStreams[0].LogStreamName);
-                FilterLogEventsResponse filterLogEventsResponse = Client.
-                    FilterLogEventsAsync(new FilterLogEventsRequest
-                    {
-                        FilterPattern = filterPattern,
-                        LogGroupName = logGroupName,
-                        LogStreamNames = logStreamNames
+                        LogGroupNamePrefix = logGroupName
                     }).Result;
 
-                return filterLogEventsResponse.Events.Count > 0;
-            }
-            else
-            {
-                return false;
+                    if (!(string.IsNullOrEmpty(describeLogGroupsResponse.LogGroups[0].LogGroupName)))
+                    {
+                        var response = Client.DeleteLogGroupAsync(new DeleteLogGroupRequest
+                        {
+                            LogGroupName = logGroupName
+                        }).Result;
+                    }
+                }
             }
         }
     }
